backend/pyproject.toml
[tool.poetry]
name = "fantasyfootball"
version = "0.1.0"
description = ""
authors = ["aleha2001 <alehak2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.105.0"}
mongoengine = "^0.27.0"
python-dotenv = "^1.0.0"
pre-commit = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
